[Update]: 1/2

+++ cpp03/ex02/ClapTrap.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ClapTrap.hpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seoson <seoson@student.42.fr>              +#+  +:+       +#+        */
+/*   By: Sungho <Sungho@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 16:36:15 by seoson            #+#    #+#             */
-/*   Updated: 2023/12/29 13:12:33 by seoson           ###   ########.fr       */
+/*   Updated: 2024/01/02 14:38:53 by Sungho           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@
 class ClapTrap
 {
 	protected:
-		std::string			name;
-		unsigned long long	Hit_points;
-		unsigned long long	Energy_points;
-		unsigned long long	Attack_damage;
+		std::string		name;
+		unsigned int	Hit_points;
+		unsigned int	Energy_points;
+		unsigned int	Attack_damage;
 	public:
 		ClapTrap();
 		ClapTrap(std::string name);
