+++ cpp03/ex03/ClapTrap.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ClapTrap.hpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seoson <seoson@student.42.fr>              +#+  +:+       +#+        */
+/*   By: Sungho <Sungho@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 16:36:15 by seoson            #+#    #+#             */
-/*   Updated: 2023/12/29 13:12:33 by seoson           ###   ########.fr       */
+/*   Updated: 2024/01/02 14:26:36 by Sungho           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,16 +18,16 @@
 class ClapTrap
 {
 	protected:
-		std::string			name;
-		unsigned long long	Hit_points;
-		unsigned long long	Energy_points;
-		unsigned long long	Attack_damage;
+		std::string		name;
+		unsigned int	Hit_points;
+		unsigned int	Energy_points;
+		unsigned int	Attack_damage;
 	public:
 		ClapTrap();
 		ClapTrap(std::string name);
 		~ClapTrap();
 		ClapTrap(const ClapTrap& claptrap);
-		ClapTrap	&operator=(const ClapTrap& claptrap);
+		ClapTrap&	operator=(const ClapTrap& claptrap);
 		void		attack(const std::string& target);
 		void		takeDamage(unsigned int amount);
 		void		beRepaired(unsigned int amount);
